feat(2fa): adjusting user data retrieval (#12020)
ref: MANAGER-14035
