+++ packages/manager/apps/procedures/src/App.tsx
@@ -6,8 +6,29 @@ import { RouterProvider, createHashRouter } from 'react-router-dom';
 import '@ovhcloud/ods-theme-blue-jeans';
 import queryClient from './query.client';
 import { Routes } from './routes/routes';
+import { decodeToken, extractToken } from '@/utils/token';
+import initI18n from './i18n';
+import initAuthenticationInterceptor from '@/data/authentication.interceptor';
+import initInterceptor from './data/invisible-challenge.interceptor';
+import UserProvider from '@/context/User/provider';
+import { getRedirectLoginUrl } from '@/utils/url-builder';
 
-odsSetup();
+const token = extractToken();
+const user = decodeToken(token);
+const { language: locale, subsidiary } = user || {};
+
+// If we don't have a user or the "session" is expired we redirect the user on the login page
+if (!user || user.exp * 1000 < Date.now()) {
+  const redirectUrl = getRedirectLoginUrl(user);
+  window.location.assign(redirectUrl);
+} else {
+
+  initI18n(locale || 'en_GB', [locale || 'en_GB'], subsidiary);
+  initAuthenticationInterceptor(token);
+  initInterceptor();
+
+  odsSetup();
+}
 
 const router = createHashRouter(Routes);
 
@@ -16,8 +37,10 @@ const Router = () => <RouterProvider router={router} />;
 export default function App() {
   return (
     <QueryClientProvider client={queryClient}>
-      <Router />
-      <ReactQueryDevtools initialIsOpen={false} />
+      <UserProvider user={user}>
+        <Router />
+        <ReactQueryDevtools initialIsOpen={false} />
+      </UserProvider>
     </QueryClientProvider>
   );
 }
